fix: resolve Iceberg int vs long type mismatch on CSV ingest
- tools.py: cast PyArrow columns to match Iceberg table schema before append()
  This fixes 'Mismatch in fields: optional int vs optional long' error
- tools.py: add LongType import and map 'long'/'int64' to LongType() (64-bit)
- main.py: map pandas int64 -> 'long' in CSV schema detection so new tables
  use LongType matching pandas default, preventing future type conflicts
- Add regression test suite (backend/tests/test_regression.py)
- Add load test suite (backend/locustfile.py) with multi-user classes
- Add large dataset generator (backend/generate_large_dataset.py)
- Add Playwright UI test suite (tests/ui/) for Chat, MCP, Ingest, Workspace tabs

+++ backend/api/main.py
@@ -609,9 +609,10 @@ async def upload_csv_endpoint(
             
         df = pd.read_csv(temp_file_path)
         
-        # Auto-detect schema
+        # Auto-detect schema — use "long" (int64) to match pandas/PyArrow defaults.
+        # Iceberg IntegerType is 32-bit; LongType is 64-bit (what pandas produces).
         type_map = {
-            "int64": "integer", "int32": "integer",
+            "int64": "long", "int32": "integer",
             "float64": "double", "float32": "float",
             "bool": "boolean",
             "object": "string",
